# backend/openapi.yaml
openapi: 3.0.3
info:
  title: ACS Operator REST API
  version: "1.0.0"
  description: >-
    The operator-facing REST API served by cmd/api (design doc v3 §8,
    build plan §4). Closes the "OpenAPI spec — still deferred" gap
    tracked in tr069-acs-build-plan.md §10 — hand-written against the
    actual route table and handler request/response structs in
    cmd/api/*.go as of this pass, not generated, so it may drift from the
    code over time; treat cmd/api/main.go's route() registrations as the
    source of truth if this ever disagrees.

    Every write endpoint (POST/PUT/DELETE that mutates device state)
    returns 202 Accepted and a command_key — it only ever *queues* a job
    for cmd/acs to dispatch on the device's next CWMP session, it never
    talks to a CPE directly (design doc v3 §8.3/§19.2). Poll
    GET /api/v1/jobs/{command_key} for the outcome.

    Authentication is opt-in: unset ACS_JWT_SIGNING_SECRET and every
    route here (except the CPE-facing firmware/upload routes, which are
    always public — a CPE has no operator JWT to present) runs
    unauthenticated, matching Phases 1-5's original behavior. Set it and
    every route requires a Bearer JWT from POST /api/v1/auth/login,
    enforced by role (readonly < noc < manager < superadmin) and, for a
    curated set of higher-stakes capabilities, by a separately
    configurable role_permissions table (GET/PUT
    /api/v1/auth/role-permissions) — see the Permission enum on write
...

...
servers:
...
security:
  - bearerAuth: []
  - {}
...
paths:
...
  /api/v1/devices/import:
    post:
      tags: [Reports]
      summary: Bulk-create/update devices from an uploaded file (tenancy.manage permission, capped at 5000 rows)
      parameters:
        - name: format
          in: query
          required: true
          schema: { type: string, enum: [json, csv, xml] }
      requestBody:
        required: true
        content:
          application/json: { schema: { type: string, format: binary } }
          text/csv: { schema: { type: string, format: binary } }
          application/xml: { schema: { type: string, format: binary } }
      responses:
        "200": { description: Import result, content: { application/json: { schema: { $ref: "#/components/schemas/ImportResult" } } } }
        "400": { description: "No rows, too many rows (>5000), or unparseable body" }

  # --- Devices ---
...
  /api/v1/firmware/images:
    get:
      tags: [Firmware & Rollouts]
      summary: List uploaded firmware images
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/FirmwareImage" } } } } } } }
    post:
      tags: [Firmware & Rollouts]
      summary: Upload a firmware image (firmware.manage permission, multipart)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                vendor: { type: string }
                model: { type: string }
                version: { type: string }
                channel: { type: string }
      responses:
        "201": { description: Uploaded, content: { application/json: { schema: { $ref: "#/components/schemas/FirmwareImage" } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "From POST /api/v1/auth/login. Enforcement is opt-in — see the top-level description."
...
  schemas:
...
    FirmwareImage:
      type: object
      properties:
        id: { type: string }
        vendor: { type: string }
        model: { type: string }
        version: { type: string }
        channel: { type: string }
        filename: { type: string }
        file_size_bytes: { type: integer }
        created_at: { type: string, format: date-time }
...
    ImportResult:
      type: object
      properties:
        total_rows: { type: integer }
        succeeded: { type: integer }
        failed: { type: integer }
        results:
          type: array
          items:
            type: object
            properties: { serial_number: { type: string }, status: { type: string }, error: { type: string } }
...